Functions

Main.ViconUDP.clear_UDP_buffer — Method

clear_UDP_buffer(vicon::ViconSystem)

Helper function that clears the UDP buffer. The timestamp is set as soon as the packet is read. No timestamp is transmitted via the UDP interface. The UDP buffer must therefore be cleard before reading starts.

TODO: Is there a better variant via the Julia socket?

Main.ViconUDP.close_vicon — Method

close_vicon(vicon::ViconSystem)

takes the byte_string from the Vicon UDP socket and returns a Vector of ItemStructs

Main.ViconUDP.measure_input_frequenzy — Method

measure_input_frequenzy(byte_string::Vector{UInt8})

takes the byte_string from the Vicon UDP socket and returns a Vector of ItemStructs

Main.ViconUDP.read_vicon — Method

read_vicon(vicon::ViconSystem)

reads a UDP package and returns the selected item.

Main.ViconUDP.read_vicon — Method

read_vicon(vicon::ViconSystem)

reads a UDP package and returns a vector of items.

Main.ViconUDP.reinterpret_Vicon — Method

reinterpret_Vicon(byte_string::Vector{UInt8})

takes the byte_string from the Vicon UDP socket and returns a Vector of raw ItemStructs

Main.ViconUDP.set_world! — Method

set_world!(vicon::ViconSystem, ItemName::String)

Set transformation from Vicon System to World based on a given item.

Main.ViconUDP.set_world! — Method

set_world!(vicon::ViconSystem, x_Vicon2World_m::Vector, R_Vicon2World::Matrix)

Set manual transformation from Vicon System to World.

Main.ViconUDP.start_async_read — Method

start_async_read(vicon::ViconSystem, ItemName::String, buffer_size = 10)

Starts a asyncron read from Vicon System. Returns a function that returns the latest item. buffer_size is the size of the Channel sending data from the async task.

Main.ViconUDP.transform_to_world — Method

transform_to_world(vicon::ViconSystem, Item::ItemStruct)

Set transformation from Vicon System to World based on a given item.

Main.ViconUDP.ItemStruct — Type

ItemStruct

Struct for the Vicon UDP data

Main.ViconUDP.ViconSystem — Type

ViconSystem

Struct for the Vicon System. Opens the UDP socket and binds it to the given ip and port.